JHEMMS-15 More code restructuring
